django/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1
django-redis==5.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.8
sqlparse==0.5.1
urllib3==2.2.2
gunicorn==21.2.0
requests==2.32.3
pillow==11.0.0
whitenoise==6.8.2
